Split UpdateHearingService into separate services
