+++ BookingsApi/BookingsApi/ConfigureServicesExtensions.cs
@@ -125,7 +125,8 @@ namespace BookingsApi
             services.AddScoped<IHearingParticipantService, HearingParticipantService>();
             services.AddScoped<IParticipantAddedToHearingAsynchronousProcess, ParticipantAddedToHearingAsynchronousProcess>();
             services.AddScoped<INewJudiciaryAddedAsynchronousProcesses, NewJudiciaryAddedAsynchronousProcesses>();
-            services.AddScoped<IUpdateHearingService, UpdateHearingService>();
+            services.AddScoped<IEndpointService, EndpointService>();
+            services.AddScoped<IJudiciaryParticipantService, JudiciaryParticipantService>();
             RegisterCommandHandlers(services);
             RegisterQueryHandlers(services);
